Added footer in albums

+++ albums.php
@@ -4,18 +4,23 @@
 	<meta charset="UTF-8">
 	<title>Albums</title>
 	<!-- Latest compiled and minified CSS -->
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+	<link rel="stylesheet" href="css/bootstrap.min.css">
 	<!-- Optional theme -->
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
+	<link rel="stylesheet" href="css/bootstrap-theme.min.css">
 
 </head>
 <body>
 	<h1>Albums</h1>
-	<a href="AlbumDetails.php" >Add a new Album</a>
+	<?php
+	session_start();
+
+	if (!empty($_SESSION['email']))
+		echo '<a href="AlbumDetails.php" >Add a new Album</a>';
+	?>
 
 	<?php
 		//Step 1 - connect to the db
-		$conn = new PDO('mysql:host=aws.computerstudi.es;dbname=gc200340662', 'gc200340662','uozYSDupBu');
+		require_once('db.php');
 		//Step 2 - create a SQL command
 		$sql = "SELECT * FROM albums";
 		//Step 3 - prepare the command
@@ -31,20 +36,26 @@
 			<tr><th>Title</th>
 				<th>Year</th>
 				<th>Artist</th>
-				<th>Genre</th>
-				<th>Edit</th>
-				<th>Delete</th></tr>';
+				<th>Genre</th>';
+
+		if (!empty($_SESSION['email'])) {
+			echo 	'<th>Edit</th>
+					<th>Delete</th>';
+		}
+		echo '</tr>';
+
 		foreach ($albums as $album) {
 			echo '<tr>	<td>'.$album['title'].'</td>
 						<td>'.$album['year'].'</td>
 						<td>'.$album['artist'].'</td>
-						<td>'.$album['genre'].'</td>
-						<td><a href="AlbumDetails.php?albumID='.$album['albumID'].'" class="btn btn-primary">Edit</a></td>
-						<td><a href="deleteAlbum.php?albumID='.$album['albumID'].'" class="btn btn-danger">Delete</a></td>';
+						<td>'.$album['genre'].'</td>';
+			if (!empty($_SESSION['email'])) {
+				echo 	'<td><a href="AlbumDetails.php?albumID='.$album['albumID'].'" class="btn btn-primary">Edit</a></td>
+						<td><a href="deleteAlbum.php?albumID='.$album['albumID'].'" class="btn btn-danger confirmation">Delete</a></td>';
+			}
+			echo '</tr>';
 		}
 		echo '</table>';
+
+		include_once('footer.php');
 	?>
-</body>
-<!-- Latest compiled and minified JavaScript -->
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-</html>
